finished update. quite some time improvements
